move query/users to query_users crate
