+++ query/users/src/auth/mod.rs
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub(crate) mod auth_mgr;
+pub mod auth_mgr;
 mod jwt;
