+++ query/users/src/role_mgr.rs
@@ -19,7 +19,7 @@ use common_meta_types::RoleIdentity;
 use common_meta_types::RoleInfo;
 use common_meta_types::UserPrivilegeSet;
 
-use crate::users::UserApiProvider;
+use crate::UserApiProvider;
 
 impl UserApiProvider {
     // Get one role from by tenant.
