+++ query/users/src/user_stage.rs
@@ -16,7 +16,7 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_types::UserStageInfo;
 
-use crate::users::UserApiProvider;
+use crate::UserApiProvider;
 
 /// user stage operations.
 impl UserApiProvider {
